requirements.txt
langchain
tweepy
feedparser
newspaper3k
openai
python-dotenv